Merge: Merged branch main into branch feat/homepage, adjusting test files temporarily in order to run them all together
