+++ src/database/colors.js
@@ -1,7 +1,7 @@
 import connection from "./connection.js";
 
 async function add({colorName, hexCode}) {
-    const result = await connection.query(`INSERT INTO colors (name, hex_code) VALUES ($1, $2) RETURNING *;`,[colorName, hexCode])
+    const result = await connection.query(`INSERT INTO colors (name, hex_code) VALUES ($1, $2) RETURNING *;`, [colorName, hexCode]);
     return result.rows[0];
 }
